Second commit

+++ src/main/java/com/example/libreria/repo/LibroRepo.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 
 
 public interface LibroRepo extends JpaRepository<Libro, Long> {
-    void deleteLibroById(int id);
+    void deleteLibroById(Long id);
 
     Optional<Libro> findLibroById(Long id);
     List<Libro> findLibroByAutorDni(String dni);
